refactor: Move to ChaosInitiative.ScriptSystem. namespace and remove unnecessary imports
