+++ ChaosInitiative.ScriptSystem.Core/AssemblyAttributes.cs
@@ -1,0 +1,4 @@
+using System.Runtime.CompilerServices;
+
+// Make our internal classes visible to the test assembly
+[assembly:InternalsVisibleTo("ChaosInitiative.ChaosInitiative.ScriptSystem.Core.Tests")]
