+++ ChaosInitiative.ScriptSystem.Core/Engine/Entities/BaseEntity.cs
@@ -1,0 +1,7 @@
+namespace ChaosInitiative.ScriptSystem.Core.Engine.Entities
+{
+    public class BaseEntity
+    {
+
+    }
+}
