+++ ChaosInitiative.ScriptSystem.Core/Engine/Entities/BasePlayer.cs
@@ -1,0 +1,7 @@
+namespace ChaosInitiative.ScriptSystem.Core.Engine.Entities
+{
+    public class BasePlayer
+    {
+
+    }
+}
