+++ ChaosInitiative.ScriptSystem.Core/IO/IFileWatcher.cs
@@ -1,0 +1,7 @@
+namespace ChaosInitiative.ScriptSystem.Core.IO
+{
+    public interface IFileWatcher
+    {
+        
+    }
+}
