+++ ChaosInitiative.ScriptSystem.Core/IO/Native/NativeFileWatcher.cs
@@ -1,0 +1,7 @@
+namespace ChaosInitiative.ScriptSystem.Core.IO.Native
+{
+    internal class NativeFileWatcher : IFileWatcher
+    {
+        
+    }
+}
