+++ ChaosInitiative.ScriptSystem.Core/Modules/Dependencies/DependencyManager.cs
@@ -1,0 +1,10 @@
+namespace ChaosInitiative.ScriptSystem.Core.Modules.Dependencies
+{
+    /// <summary>
+    /// Resolves inter-module dependencies.
+    /// </summary>
+    internal class DependencyManager
+    {
+
+    }
+}
