+++ ChaosInitiative.ScriptSystem.Core/Modules/IScriptModuleRepository.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace ChaosInitiative.ScriptSystem.Core.Modules
+{
+    public interface IScriptModuleRepository
+    {
+        Task<IEnumerable<ScriptModule>> EnumerateModulesAsync();
+    }
+}
