+++ ChaosInitiative.ScriptSystem.Core/Modules/ScriptModuleLoadState.cs
@@ -1,0 +1,9 @@
+namespace ChaosInitiative.ScriptSystem.Core.Modules
+{
+    public enum ScriptModuleLoadState
+    {
+        None,
+        Compiling,
+        Loading
+    }
+}
