+++ ChaosInitiative.ScriptSystem.RuntimeTest/ChaosInitiative.ScriptSystem.RuntimeTest.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\ChaosInitiative.ScriptSystem.Core\ChaosInitiative.ScriptSystem.Core.csproj" />
+  </ItemGroup>
+  
+</Project>
