--- ScriptSystem.Core/AssemblyAttributes.cs
@@ -1,4 +1,0 @@
-using System.Runtime.CompilerServices;
-
-// Make our internal classes visible to the test assembly
-[assembly:InternalsVisibleTo("ScriptSystem.Core.Tests")]
