--- ScriptSystem.Core/Engine/Entities/BaseEntity.cs
@@ -1,7 +1,0 @@
-namespace ScriptSystem.Core.Engine.Entities
-{
-    public class BaseEntity
-    {
-
-    }
-}
