--- ScriptSystem.Core/IO/IFileWatcher.cs
@@ -1,7 +1,0 @@
-namespace ScriptSystem.Core.IO
-{
-    public interface IFileWatcher
-    {
-        
-    }
-}
