--- ScriptSystem.Core/IO/Native/NativeFileWatcher.cs
@@ -1,10 +1,0 @@
-using System.IO;
-using System.Threading.Tasks;
-
-namespace ScriptSystem.Core.IO.Native
-{
-    internal class NativeFileWatcher : IFileWatcher
-    {
-        
-    }
-}
