--- ScriptSystem.Core/Modules/Dependencies/DependencyManager.cs
@@ -1,12 +1,0 @@
-using System;
-
-namespace ScriptSystem.Core.Modules.Dependencies
-{
-    /// <summary>
-    /// Resolves inter-module dependencies.
-    /// </summary>
-    internal class DependencyManager
-    {
-
-    }
-}
