--- ScriptSystem.Core/Modules/IScriptModuleRepository.cs
@@ -1,10 +1,0 @@
-using System.Threading.Tasks;
-using System.Collections.Generic;
-
-namespace ScriptSystem.Core.Modules
-{
-    public interface IScriptModuleRepository
-    {
-        Task<IEnumerable<ScriptModule>> EnumerateModulesAsync();
-    }
-}
