--- ScriptSystem.Core/Modules/ScriptModuleLoadState.cs
@@ -1,6 +1,0 @@
-public enum ScriptModuleLoadState
-{
-    None,
-    Compiling,
-    Loading
-}
