--- ScriptSystem.RuntimeTest/Program.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.IO;
-using System.Threading.Tasks;
-
-using ScriptSystem.Core.IO.Native;
-
-namespace ScriptSystem.RuntimeTest
-{
-    class Program
-    {
-        static async Task Main(string[] args)
-        {
-            var fileSystem = new NativeFileSystem(".");
-            using (var stream = fileSystem.Open("Program.cs", FileMode.Open, FileAccess.Read))
-            {
-                using (var sr = new StreamReader(stream))
-                {
-                    Console.Write(sr.ReadToEnd());
-                }
-            }
-
-            //var manager = new ScriptModuleManager();
-            //await manager.ReloadAllAsync();
-        }
-    }
-}
