+++ ScriptSystem.sln
@@ -3,11 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26124.0
 MinimumVisualStudioVersion = 15.0.26124.0
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ScriptSystem.Core", "ScriptSystem.Core\ScriptSystem.Core.csproj", "{56AB402F-A94B-41B5-BBC2-0BC1EA1DEE2B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ChaosInitiative.ScriptSystem.Core", "ChaosInitiative.ScriptSystem.Core\ChaosInitiative.ScriptSystem.Core.csproj", "{56AB402F-A94B-41B5-BBC2-0BC1EA1DEE2B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ScriptSystem.Core.Tests", "ScriptSystem.Core.Tests\ScriptSystem.Core.Tests.csproj", "{9F4AA31F-1932-457A-9EFB-566715F43278}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ChaosInitiative.ScriptSystem.Core.Tests", "ChaosInitiative.ScriptSystem.Core.Tests\ChaosInitiative.ScriptSystem.Core.Tests.csproj", "{9F4AA31F-1932-457A-9EFB-566715F43278}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ScriptSystem.RuntimeTest", "ScriptSystem.RuntimeTest\ScriptSystem.RuntimeTest.csproj", "{5D310595-EB66-4A19-8C68-0223A0C1A1B9}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ChaosInitiative.ScriptSystem.RuntimeTest", "ChaosInitiative.ScriptSystem.RuntimeTest\ChaosInitiative.ScriptSystem.RuntimeTest.csproj", "{5D310595-EB66-4A19-8C68-0223A0C1A1B9}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
